Adding new metric pdcsi_operation_errors to fetch error count
